V2.0.1-Beta-Wordpress-install-Advanced-Setup
V2.0.1-Beta-Wordpress-install-Advanced-Setup
Added Backup
Advance help
Prefix
Validation checks
User friendly
And more.

TO DO
Needs to be tested in a few environments. 
Need to test in multiple Linux Operating systems. Ubuntu Debian.
Needs to be able to ask what PHP to install. 
Needs to be away to identify the system and based on that do installs, 
Possibly refactor the script

Possibly add in the future, A list of plugin you can select and download and install into WordPress. 
If so research to be done on whats best out there, and if its free or paid. 
if paid, let the user login via terminal or download manually into root of script.